Рефакторинг глобального контекста
